refactor(editor): table overlay residual 분리
